+++ BasketTestLib/BasketTestLib/Models/Gloves.cs
@@ -2,7 +2,7 @@
 {
     public class Gloves : Product
     {
-        public Gloves(float unitPrice) : base(unitPrice)
+        public Gloves(decimal unitPrice) : base(unitPrice)
         {
 
         }
